+++ src/components/auth/ready.js
@@ -1,5 +1,5 @@
-import retryImport from '../../../lib/utils/retryImport'
-import logger from '../../../lib/logger/pino-logger'
+import retryImport from '../../lib/utils/retryImport'
+import logger from '../../lib/logger/pino-logger'
 
 const log = logger.child({ from: 'Ready' })
